Refactor modals code a bit, and create a AttemptResultInput component. The AttemptResultInput component still needs to be implemented, but I think it provides the right abstraction.
